Create README.md
This program can be run to get solution for a particular sudoku. You just have to enter the required numbers to get complete sudoku. If more than 1 answer are possible(may happen if numbers given are less) then any 1 possible answer can be printed. The GUI was developed using tkinter in python.